Now the script is fully recursive :D

+++ autogen.sh
@@ -1,20 +1,29 @@
 #! /bin/sh
 
-# Call sub-autogens
 SUBDIRS=
 
 for dir in $SUBDIRS
 do
     cd $dir
-    mkdir -p m4
-    rm -rf autom4te.cache
+    ./autogen.sh bare
     cd ..
 done
 
-# Generates project files autotools.
-echo "===== $(pwd)/autogen.sh ====="
-mkdir -p m4
-set -e
-autoreconf --install --force
-rm -rf autom4te.cache
+if [ $# == 0 ]
+then
+    mkdir -p m4
+    set -e
+    autoreconf --install --force
+    rm -rf autom4te.cache
+else
+    if [ $1 == "bare" ]
+    then
+        mkdir -p m4
+        rm -rf autom4te.cache
+    else
+        echo "Usage: $0 [bare]"
+    fi
+fi
+
+
